[NFC] Move hlsl::DiagnoseTranslationUnit to separate file.
Since DiagnoseTranslationUnit is after Sema, it doesn't need to be part of SemaHLSL.
This is in preparation for moving DiagnoseHLSLMethodCall to hlsl::DiagnoseTranslationUnit for access to the call graph.

For https://github.com/microsoft/DirectXShaderCompiler/issues/5855
