+++ tools/clang/include/clang/Sema/SemaHLSL.h
@@ -69,11 +69,6 @@ void DiagnosePackingOffset(clang::Sema *self, clang::SourceLocation loc,
 void DiagnoseRegisterType(clang::Sema *self, clang::SourceLocation loc,
                           clang::QualType type, char registerType);
 
-clang::FunctionDecl *ValidateNoRecursion(clang::Sema *self,
-                                         clang::FunctionDecl *FD);
-
-void ValidateNoRecursionInTranslationUnit(clang::Sema *self);
-
 void DiagnoseTranslationUnit(clang::Sema *self);
 
 void DiagnoseUnusualAnnotationsForHLSL(
